requirements.txt
# Core dependencies
flask==3.0.0
django==5.0.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9  # Changed from psycopg2
pymongo==4.6.1

# ML/AI
transformers==4.36.2

# Development
pre-commit==3.6.0

# Add explicit versions for pydantic
pydantic==2.4.2  
typing-extensions>=4.6.1

# Auth & Security
PyJWT==2.8.0

mongomock==4.1.2  # For MongoDB mocking
pytest-mock==3.10.0  # For general mocking utilities
pytest-cov==4.1.0  # For coverage reporting
pytest-xdist==3.3.1  # For parallel test execution